Allow for SQL commands and not just queries
The current wrapper only allows (expects) queries, not commands, so TRUNCATE, DROP, CREATE are not available.

Added a ``stmt->execute(query)`` call to fix.

This pull has renamed the excising 'execute' to 'query' and might not be preferable, required excising code to be changed if updated with this. It might be better to keep the 'execute' name as is and create the new as 'command' or something?

Also the header file will need updating. ;-)